Training utilities refactored. Just need to modify imports in the harnesses and tests.
